First step in eliminating selfie mode
No longer on the menu, and now the camera stays fixed, and you
can move around to frame things before it takes the snapshot.

I will figure out how to do this with the existing independent
camera mode next, but for now it appears you cannot move when
in that mode.  More to come.